Put the FutureAdapter in the PlatformRef impl
